Remove top level setup.py
