+++ mlcommons_box/setup.py
@@ -1,4 +1,4 @@
-from setuptools import find_packages, setup
+from setuptools import find_packages, setup, Command
 import os
 
 # get the dependencies and installs
@@ -8,6 +8,16 @@ with open("requirements.txt", "r", encoding="utf-8") as f:
         req = line.split("#", 1)[0].strip()
         if req and not req.startswith("--"):
             requires.append(req)
+
+class clean(Command):
+    """Custom clean command."""
+    user_options = []
+    def initialize_options(self):
+        pass
+    def finalize_options(self):
+        pass
+    def run(self):
+        os.system('rm -vrf ./build ./dist ./*.egg-info')
 
 def schemas():
     paths = []
@@ -22,11 +32,16 @@ setup(
     name="mlcommons_box",
     version="0.1",
     packages=find_packages(exclude=["tests"]),
+    license="Apache 2.0",
     entry_points='''
         [console_scripts]
         mlcommons_box=mlcommons_box.main:cli
     ''',
     install_requires=requires,
+    python_requires='>=3.6',
     package_data={"": extra_files},
     extras_require={},
+    cmdclass={
+        'clean': clean,
+    }
 )
