--- requirements.txt
@@ -1,3 +1,0 @@
-mlspeclib
-docker
-pyyaml
